feat: MyInfoChange 컴포넌트 닉네임 체크 완성
#8
